API and Ajax: just basic cleaning and making rubocop happy. ObserverController: first bady step toward splitting.
