+++ app/classes/api/parse_external_site.rb
@@ -6,10 +6,10 @@ class API
     declare_parameter(key, :external_site, args)
     str = get_param(key)
     return args[:default] unless str
-    fail BadParameterValue.new(str, :external_site) if str.blank?
+    raise BadParameterValue.new(str, :external_site) if str.blank?
     val = try_parsing_id(str, ExternalSite)
     val ||= ExternalSite.find_by_name(str)
-    fail ObjectNotFoundByString.new(str, ExternalSite) unless val
+    raise ObjectNotFoundByString.new(str, ExternalSite) unless val
     val
   end
 end
