+++ app/controllers/ajax_controller/api_key.rb
@@ -1,4 +1,5 @@
 # encoding: utf-8
+# see ajax_controller.rb
 class AjaxController
   # Activate mode: sets verified field of given ApiKey, returns nothing.
   # Edit mode: sets notes field of given ApiKey, returns new value.
