+++ app/controllers/ajax_controller/export.rb
@@ -1,4 +1,5 @@
 # encoding: utf-8
+# see ajax_controller.rb
 class AjaxController
   # Mark an object for export. Renders updated export controls.
   # type::  Type of object.
