+++ app/controllers/ajax_controller/external_link.rb
@@ -1,4 +1,5 @@
 # encoding: utf-8
+# see ajax_controller.rb
 class AjaxController
   # Add, edit or remove external link associated with an observation.
   # Returns link id on success.
