+++ app/controllers/ajax_controller/geocode.rb
@@ -1,4 +1,5 @@
 # encoding: utf-8
+# see ajax_controller.rb
 class AjaxController
   require_dependency "geocoder"
