+++ app/controllers/ajax_controller/old_translation.rb
@@ -1,4 +1,5 @@
 # encoding: utf-8
+# see ajax_controller.rb
 class AjaxController
   # Return an old TranslationString by version id.
   def old_translation
