+++ app/controllers/observer_controller/theme_controller.rb
@@ -1,0 +1,7 @@
+# encoding: utf-8
+# TODO: move this into a new ThemeController
+class ObserverController
+  # Info on color themes.
+  def color_themes # :nologin:
+  end
+end
